# Clean up interpreter primops