--- a/unison-runtime/unison-runtime.cabal
+++ b/unison-runtime/unison-runtime.cabal
@@ -52,6 +52,8 @@ library
       Unison.Runtime.Interface
       Unison.Runtime.IOSource
       Unison.Runtime.Machine
+      Unison.Runtime.Machine.Primops
+      Unison.Runtime.Machine.Types
       Unison.Runtime.MCode
       Unison.Runtime.MCode.Serialize
       Unison.Runtime.Pattern