This PR refactors the primop parts of the interpreter to improve things in a few ways.

1. The primop types have been coalesced into just `Prim1` and `Prim2`. They used to be broken up by unboxed vs. boxed, but that distinction has mostly been removed. This reduces the number of constructors necessary in the instruction type.
2. Most primop implementations have been moved to a dedicated file, and have been factored into an 'action' and a wrapper. This is in anticipation of wrapping the action in different ways (though that doesn't exist yet). I also was able to remove several special cased primops, so only a few of those remain in the main interpreter file.
3. A few primops have been changed to use `ForeignConvention` for their results rather than stack encoding, which enables some speedups as well as making the implementation easier.

Ideally this is a step toward having primops applied to literals, but the whole changeset for that was slower. This is a significant chunk that has (I believe) no performance regressions, and is in fact faster on some things.

Benchmark results with noticeable improvement (others are about the same; probably just random variation):

```
Decode Nat
198ns -> 182ns

List.foldLeft
2.615312ms -> 1.901392ms

Json parsing (per document)
302.499µs -> 284.651µs

Count to N (per element)
113ns -> 107ns

Count to 1000
114.441µs -> 106.719µs

Mutate a Ref 1000 times
176.644µs -> 166.901µs

CAS an IO.ref 1000 times
202.658µs -> 193.969µs

List.range (per element)
259ns -> 249ns

Set.fromList (range 0 1000)
1.764502ms -> 1.541366ms

Map.fromList (range 0 1000)
1.237594ms -> 1.051414ms

List.at (1k element list)
222ns -> 193ns
```

Pretty sure these are due to the better calling conventions/wrappers.